backend/requirements.txt
Flask
Pillow
nltk
textblob
